pyproject.toml
[project]
name = "MEDUSA-PLATFORM"
version = "0.1.1"
requires-python = ">=3.13.12"
dependencies = [
    "bson==0.5.10",
    "buildtools>=1.0.6",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "colorama==0.4.6",
    "contourpy==1.3.3",
    "cryptography==46.0.3",
    "cycler==0.12.1",
    "dill==0.4.0",
    "fonttools==4.61.0",
    "h5py==3.15.1",
    "idna==3.11",
    "ipykernel>=7.1.0",
    "jinja2==3.1.6",
    "joblib==1.5.2",
    "kiwisolver==1.4.9",
    "markupsafe==3.0.3",
    "matplotlib==3.10.7",
    "medusa-kernel>=1.4,<1.5",
    "numpy==2.3.5",
    "packaging==25.0",
    "pandas==2.3.3",
    "patsy==1.0.2",
    "pillow==12.0.0",
    "pycparser==2.23",
    "pylsl==1.17.6",
    "pyparsing==3.2.5",
    "pyqt5-qt5==5.15.2",
    "pyside6==6.10.1",
    "pyside6-addons==6.10.1",
    "pyside6-essentials==6.10.1",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pywavelets==1.9.0",
    "requests==2.32.5",
    "scikit-learn==1.7.2",
    "scipy==1.16.3",
    "setuptools>=80.10.2",
    "shiboken6==6.10.1",
    "six==1.17.0",
    "statsmodels==0.14.5",
    "threadpoolctl==3.6.0",
    "tqdm==4.67.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "wheel>=0.46.3",
]

[dependency-groups]
dev = [
    "cython>=3.2.4",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    # "pyqtinspect=^0.3.6",
    "pyinstaller>=6.18.0",
    "setuptools>=80.10.2",
    "wheel>=0.46.3",
]

[tool.uv.sources]
# Use local medusa-kernel when developing in MedusaRoot (so kernel changes are picked up)
medusa-kernel = { path = "../medusa-kernel", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/medusa_platform"]

# [build-system]
# requires = ["setuptools>=61.0"]
# build-backend = "setuptools.build_meta"

# [tool.setuptools]
# package-dir = { "" = "src" }
# packages = { find = { where = ["src"] } }
# py-modules = ["accounts_manager", "app_manager", "constants", "exceptions", "main", "resources", "updates_manager", "user_session", "utils"]
